requirements.txt
optuna
scikit-learn
pandas
seaborn
plotly==5.24.1
torchvision
fastapi
uvicorn
python-multipart
redis
